+++ MainCore/Helper/Implementations/Base/HeroHelper.cs
@@ -1,0 +1,48 @@
+﻿using FluentResults;
+using MainCore.Enums;
+using MainCore.Helper.Interface;
+using MainCore.Parser.Interface;
+using MainCore.Services.Interface;
+using OpenQA.Selenium;
+
+namespace MainCore.Helper.Implementations.Base
+{
+    public abstract class HeroHelper : IHeroHelper
+    {
+        protected readonly IChromeManager _chromeManager;
+        protected readonly IHeroSectionParser _heroSectionParser;
+        protected readonly INavigateHelper _navigateHelper;
+
+        public HeroHelper(IChromeManager chromeManager, IHeroSectionParser heroSectionParser, INavigateHelper navigateHelper)
+        {
+            _chromeManager = chromeManager;
+            _heroSectionParser = heroSectionParser;
+            _navigateHelper = navigateHelper;
+        }
+
+        public abstract Result ClickItem(int accountId, HeroItemEnums item);
+
+        public Result EnterAmount(int accountId, int amount)
+        {
+            var chromeBrowser = _chromeManager.Get(accountId);
+            var doc = chromeBrowser.GetHtml();
+            var amountBox = _heroSectionParser.GetAmountBox(doc);
+            if (amountBox is null)
+            {
+                return Result.Fail("Cannot find amount box");
+            }
+            var chrome = chromeBrowser.GetChrome();
+            var amountInputs = chrome.FindElements(By.XPath(amountBox.XPath));
+            if (amountInputs.Count == 0)
+            {
+                return Result.Fail("Cannot find amount box");
+            }
+            amountInputs[0].SendKeys(Keys.Home);
+            amountInputs[0].SendKeys(Keys.Shift + Keys.End);
+            amountInputs[0].SendKeys(amount.ToString());
+            return Result.Ok();
+        }
+
+        public abstract Result Confirm(int accountId);
+    }
+}
