+++ MainCore/Helper/Implementations/Base/UpdateHelper.cs
@@ -3,27 +3,27 @@ using MainCore.Enums;
 using MainCore.Errors;
 using MainCore.Helper.Interface;
 using MainCore.Models.Database;
+using MainCore.Parser.Interface;
 using MainCore.Services.Interface;
 using Microsoft.EntityFrameworkCore;
-using MainCore.Parser.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace MainCore.Helper.Implementations
+namespace MainCore.Helper.Implementations.Base
 {
-    public class UpdateHelper : IUpdateHelper
+    public abstract class UpdateHelper : IUpdateHelper
     {
-        private readonly IVillageCurrentlyBuildingParser _villageCurrentlyBuildingParser;
-        private readonly IChromeManager _chromeManager;
-        private readonly IDbContextFactory<AppDbContext> _contextFactory;
-        private readonly IVillageFieldParser _villageFieldParser;
-        private readonly IVillageInfrastructureParser _villageInfrastructureParser;
-        private readonly IStockBarParser _stockBarParser;
-        private readonly IHeroSectionParser _heroSectionParser;
-        private readonly ISubTabParser _subTabParser;
-        private readonly IFarmListParser _farmListParser;
-        private readonly IEventManager _eventManager;
+        protected readonly IVillageCurrentlyBuildingParser _villageCurrentlyBuildingParser;
+        protected readonly IChromeManager _chromeManager;
+        protected readonly IDbContextFactory<AppDbContext> _contextFactory;
+        protected readonly IVillageFieldParser _villageFieldParser;
+        protected readonly IVillageInfrastructureParser _villageInfrastructureParser;
+        protected readonly IStockBarParser _stockBarParser;
+        protected readonly IHeroSectionParser _heroSectionParser;
+        protected readonly ISubTabParser _subTabParser;
+        protected readonly IFarmListParser _farmListParser;
+        protected readonly IEventManager _eventManager;
 
         public UpdateHelper(IVillageCurrentlyBuildingParser villageCurrentlyBuildingParser, IChromeManager chromeManager, IDbContextFactory<AppDbContext> contextFactory, IVillageFieldParser villageFieldParser, IVillageInfrastructureParser villageInfrastructureParser, IStockBarParser stockBarParser, ISubTabParser subTabParser, IHeroSectionParser heroSectionParser, IFarmListParser farmListParser, IEventManager eventManager)
         {
@@ -167,90 +167,7 @@ namespace MainCore.Helper.Implementations
             return Result.Ok();
         }
 
-        public Result UpdateDorf2(int accountId, int villageId)
-        {
-            var chromeBrowser = _chromeManager.Get(accountId);
-            var html = chromeBrowser.GetHtml();
-            var buildingNodes = _villageInfrastructureParser.GetNodes(html);
-            using var context = _contextFactory.CreateDbContext();
-            foreach (var buildingNode in buildingNodes)
-            {
-                var id = _villageInfrastructureParser.GetId(buildingNode);
-                var building = context.VillagesBuildings.Find(villageId, id);
-                var level = _villageInfrastructureParser.GetLevel(buildingNode);
-                int type = 0;
-                if (VersionDetector.IsTravianOfficial() && id == 26)
-                {
-                    type = (int)BuildingEnums.MainBuilding;
-                }
-                else if (id == 39)
-                {
-                    type = (int)BuildingEnums.RallyPoint;
-                }
-                else if (id == 40)
-                {
-                    var tribe = context.AccountsInfo.Find(accountId).Tribe;
-
-                    var wall = tribe.GetTribesWall();
-                    type = (int)wall;
-                }
-                else
-                {
-                    type = _villageInfrastructureParser.GetBuildingType(buildingNode);
-                }
-                var isUnderConstruction = _villageInfrastructureParser.IsUnderConstruction(buildingNode);
-                if (building is null)
-                {
-                    context.VillagesBuildings.Add(new()
-                    {
-                        VillageId = villageId,
-                        Id = id,
-                        Level = level,
-                        Type = (BuildingEnums)type,
-                        IsUnderConstruction = isUnderConstruction,
-                    });
-                }
-                else
-                {
-                    building.Level = level;
-                    building.Type = (BuildingEnums)type;
-                    building.IsUnderConstruction = isUnderConstruction;
-                }
-            }
-            context.SaveChanges();
-            var currentlyBuilding = context.VillagesCurrentlyBuildings.Where(x => x.VillageId == villageId).ToList();
-            if (currentlyBuilding.Count > 0)
-            {
-                foreach (var building in currentlyBuilding)
-                {
-                    var build = context.VillagesBuildings.FirstOrDefault(x => x.IsUnderConstruction && x.Type == building.Type && x.Level - building.Level < 3);
-                    if (build is null) continue;
-                    building.Location = build.Id;
-                }
-                context.SaveChanges();
-            }
-
-            var updateTime = context.VillagesUpdateTime.Find(villageId);
-            if (updateTime is null)
-            {
-                updateTime = new()
-                {
-                    VillageId = villageId,
-                    Dorf2 = DateTime.Now
-                };
-
-                context.VillagesUpdateTime.Add(updateTime);
-            }
-            else
-            {
-                updateTime.Dorf2 = DateTime.Now;
-            }
-            context.SaveChanges();
-
-            _eventManager.OnVillageCurrentUpdate(villageId);
-            _eventManager.OnVillageBuildsUpdate(villageId);
-            return Result.Ok();
-        }
+        public abstract Result UpdateDorf2(int accountId, int villageId);
 
         public Result UpdateResource(int accountId, int villageId)
         {
