+++ MainCore/Helper/Implementations/TTWars/AccessHelper.cs
@@ -1,0 +1,6 @@
+﻿namespace MainCore.Helper.Implementations.TTWars
+{
+    public class AccessHelper : Base.AccessHelper
+    {
+    }
+}
