+++ MainCore/Helper/Implementations/TTWars/BuildingsHelper.cs
@@ -1,0 +1,12 @@
+﻿using MainCore.Services.Interface;
+using Microsoft.EntityFrameworkCore;
+
+namespace MainCore.Helper.Implementations.TTWars
+{
+    public class BuildingsHelper : Base.BuildingsHelper
+    {
+        public BuildingsHelper(IDbContextFactory<AppDbContext> contextFactory, IPlanManager planManager) : base(contextFactory, planManager)
+        {
+        }
+    }
+}
