+++ MainCore/Helper/Implementations/TTWars/GithubHelper.cs
@@ -1,0 +1,6 @@
+﻿namespace MainCore.Helper.Implementations.TTWars
+{
+    public class GithubHelper : Base.GithubHelper
+    {
+    }
+}
