+++ MainCore/Helper/Implementations/TravianOfficial/AccessHelper.cs
@@ -1,0 +1,6 @@
+﻿namespace MainCore.Helper.Implementations.TravianOfficial
+{
+    public class AccessHelper : Base.AccessHelper
+    {
+    }
+}
