+++ MainCore/Helper/Implementations/TravianOfficial/GithubHelper.cs
@@ -1,0 +1,6 @@
+﻿namespace MainCore.Helper.Implementations.TravianOfficial
+{
+    public class GithubHelper : Base.GithubHelper
+    {
+    }
+}
